requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic
httpx>=0.27.0
requests==2.31.0
python-dotenv>=1.0.0
# Option 1: Use ctransformers (recommended - easier to install, works with GGUF)
ctransformers>=0.2.27

# Option 2: Alternative - llama-cpp-python (if you can get pre-built wheels)
# llama-cpp-python>=0.2.0

# Note: This implementation loads GGUF models directly in Python
# No need for llama.cpp server binary

# MongoDB and Authentication
motor>=3.3.2
pymongo>=4.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Google Maps API
googlemaps>=4.10.0

